style: polished date format and decimal points for quantity

+++ src/lib/utils/date.js
@@ -7,36 +7,35 @@
  * @returns {string} The formatted date string, or an empty string if the date is invalid.
  */
 export function formatDate(date, locale = 'en-US', options = {}) {
-  try {
-    let dateObj;
+	try {
+		let dateObj;
 
-    if (typeof date === 'string' || typeof date === 'number') {
-      dateObj = new Date(date);
-    } else if (date instanceof Date) {
-      dateObj = date;
-    } else {
-      return ''; // Invalid date input
-    }
+		if (typeof date === 'string' || typeof date === 'number') {
+			dateObj = new Date(date);
+		} else if (date instanceof Date) {
+			dateObj = date;
+		} else {
+			return ''; // Invalid date input
+		}
 
-    if (isNaN(dateObj.getTime())) {
-      return ''; // Invalid date
-    }
+		if (isNaN(dateObj.getTime())) {
+			return ''; // Invalid date
+		}
 
-    const defaultOptions = {
-      year: 'numeric',
-      month: 'long',
-      day: 'numeric',
-    };
+		const defaultOptions = {
+			year: 'numeric',
+			month: 'short',
+			day: 'numeric'
+		};
 
-    const combinedOptions = { ...defaultOptions, ...options };
+		const combinedOptions = { ...defaultOptions, ...options };
 
-    const formatter = new Intl.DateTimeFormat(locale, combinedOptions);
-    return formatter.format(dateObj);
-
-  } catch (error) {
-    console.error("Error formatting date:", error);
-    return '';
-  }
+		const formatter = new Intl.DateTimeFormat(locale, combinedOptions);
+		return formatter.format(dateObj);
+	} catch (error) {
+		console.error('Error formatting date:', error);
+		return '';
+	}
 }
 
 /**
@@ -46,16 +45,16 @@ export function formatDate(date, locale = 'en-US', options = {}) {
  * @returns {Date|null} The parsed Date object, or null if the string cannot be parsed.
  */
 export function parseDate(dateString) {
-  try {
-    const parsedDate = new Date(dateString);
+	try {
+		const parsedDate = new Date(dateString);
 
-    if (isNaN(parsedDate.getTime())) {
-      return null; // Invalid date
-    }
+		if (isNaN(parsedDate.getTime())) {
+			return null; // Invalid date
+		}
 
-    return parsedDate;
-  } catch (error) {
-    console.error("Error parsing date:", error);
-    return null;
-  }
+		return parsedDate;
+	} catch (error) {
+		console.error('Error parsing date:', error);
+		return null;
+	}
 }
